Prettier, new lang.js file
I installed Prettier, so now the code looks a little better. This meant I was kind of forced to move the long arrays of strings into a separate file, because I wanted them to be sent through prettier too, but not clog up the main file.
So now lang.js holds these arrays of strings.